[CRE-491] Used moved OffchainAggregator in chainlink (#19986)
* Minor.

* Minor.
